requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
requests
